Impliment saving a Gig

+++ GitHub/ViewModels/GigFormViewModel.cs
@@ -8,7 +8,7 @@ namespace GitHub.ViewModels
         public string Venue { get; set; }
         public string Date { get; set; }
         public string Time { get; set; }
-        public int Genre { get; set; }
+        public byte Genre { get; set; }
         // we need the list of genre from the database inorder to use as the second argument in dropdownlist
         public IEnumerable<Genre> Genres { get; set; }
     }
